Rename several scan-related interfaces.
